+++ example/mobile_app/test/bottom_navigation_test.dart
@@ -4,7 +4,6 @@ import 'package:mobile_app/main.dart';
 import 'package:mobile_app/pages/bottom_navigation_bar_page.dart';
 import 'package:mobile_app/pages/home_page.dart';
 import 'package:mobile_app/pages/notifications_page.dart';
-
 import 'helpers.dart';
 
 Future pumpBottomNavigationPage(WidgetTester tester) async {