+++ test/router_test.dart
@@ -393,4 +393,70 @@ void main() {
       ['/id2'],
     );
   });
+
+  test('makeUrl makes hash URL with null query params', () {
+    expect(
+      makeUrl(
+        pathStrategy: PathStrategy.hash,
+        path: '/new-path',
+        queryParameters: null,
+      ),
+      '#/new-path',
+    );
+  });
+
+  test('makeUrl makes hash URL with empty query params', () {
+    expect(
+      makeUrl(
+        pathStrategy: PathStrategy.hash,
+        path: '/new-path',
+        queryParameters: {},
+      ),
+      '#/new-path',
+    );
+  });
+
+  test('makeUrl makes hash URL with query params', () {
+    expect(
+      makeUrl(
+        pathStrategy: PathStrategy.hash,
+        path: '/new-path',
+        queryParameters: {'query': 'param'},
+      ),
+      '#/new-path?query=param',
+    );
+  });
+
+  test('makeUrl makes path URL with null query params', () {
+    expect(
+      makeUrl(
+        pathStrategy: PathStrategy.path,
+        path: '/new-path',
+        queryParameters: null,
+      ),
+      '/new-path',
+    );
+  });
+
+  test('makeUrl makes path URL with empty query params', () {
+    expect(
+      makeUrl(
+        pathStrategy: PathStrategy.path,
+        path: '/new-path',
+        queryParameters: {},
+      ),
+      '/new-path',
+    );
+  });
+
+  test('makeUrl makes hash URL with query params', () {
+    expect(
+      makeUrl(
+        pathStrategy: PathStrategy.path,
+        path: '/new-path',
+        queryParameters: {'query': 'param'},
+      ),
+      '/new-path?query=param',
+    );
+  });
 }